Merged PR 251187: State Transition Monitor: Making Settings Configurable
The timeouts in the state transition monitor are currently hard-coded constants, and would benefit from being settings that could be configured dynamically.

I discovered today that the suspend timeout should be lengthened, and since this was going to require a code change no matter what, I wanted to fix it once and for all by making it configurable.

Additionally, I would like to get this in before the enhancement for the provisioning state.