Merged PR 250467: Fix backplan service to proper discard stale endpoint connections of contacts
We have had for a while a behavior where the backplane services has run out of memory in several occasion durign the day. The worst scenario was seen on USW2 where in one day we have around 15 restarts. We ask the Cosmos SDK folks why suddenly the latency of upsering a contact document was so slow from millescs to several secs. Turn out the problem seems to be related of CPU spikes that make the service impossible to recover. After several tries turns out the problem was cause by our contact document in Cosmos has many 'stale' connection. I'm taking about some thousand connections accumulated over the period of months that were not properly removed. The provider implementation was not properly pushing a recoincile document based on the 'active' services beign detected and so passing back the whole blob with even more endpoint connections.
The solution was to restructure the way we consolidate the remote and local changes of a backplane. The backplane receive 'remote' contact changes from other backplanes and also local changes form the pods on the AKS. It has been design that now the 2 payloads will be treatly separatly and only merged when consolidating the data for internal consmuption or to upload back into the global backplane.
Additionl telemetry was added to allow spot this problem in the future and also to measure latency over 2 secs on the Cosmos SDK.