Merged PR 253218: feat: force upgrading to an agent compatible with the service.
feat: force upgrading to an agent compatible with the service.

OS disk for windows environments are persisted, and the vso agent inside it is not updated before the environment is resumed (unlike Linux).
This change allows vso agent to be updated before resuming the environment, using hardboot for now which is 2 minutes slower.
When a breaking change is introduced, the appsettings can be updated. When old environments are resumed, and when the agent might not be compatible with the newer service the agent will be updated.

In this case, the invitation link id will potentially break the already created environments. As the agent needs to understand how to use invitation link ids.