Merged PR 254395: chore: updating environment record with the resource records immediately after queued.
In queued mode, especially in Windows resume, the start environment continuation does not immediately update the environment record. If someone were to queue a resume, and immediately delete the environment the backend does not receive the delete for Compute because it is not in the environment record yet.

This change, updates the record immediately after the tracking resource id's are received and also after resources are ready. Any other operations that happen while the queue is in progress should do the right things especially the resource tracking for the backend.