Merged PR 254534: Improve resilency for 'unhandled exceptions' to avoid Exit 139 on the pods
The signalR services are running much more stable after many of the changes we did, but we are seeing still some restart on the pods with Exis Code 139 which could mean some of our code is not handling exceptions during background processing.
After reviewing the code it's hard to idenitfy such potential problems and so it will benefit to add try{} catch{} in critical places to log telemetry with the exception and restart some of the backgroudn services. If we see some telemetry associated with the 'undhandled' case would be able to spot the exact problem and fix it. Today we are silent on those type of errors.

This PR will also handle some fixes around refreshing app settings and providing fixes for the new Linux build.