Merged PR 253174: chore: Removing the child logger at the operation scope to avoid unncessary j...
chore: Removing the child logger at the operation scope to avoid unncessary joins in kusto.

Because a new child logger was created and passed in to the http methods, any logging that got added did not get reflected in the "request_stop" msg.
For example, to get the SkuName and the duration of the cloud environment create web call, it requires a join in kusto which shouldn't be necessary.