Merged PR 254948: Always process subscriptions regardless of their state
This change will allow reprocessing of subscriptions regardless of if their state has changed or not. During the Prod deployment the new queue message Handler for suspending environments will not go out to all regions at once thus suspension requests would be lost.

This change allows the worker to continue to request environment suspensions of subscriptions that are no longer in good standing. As the new Handler rolls out to all regions the requests will be processed.