Merged PR 254987: Do not append .git to the repo url
Do not append .git to the repo URL. This will make the behavior of the portal consistent with the extension and GitHub codespaces.

Appending `.git` at the end was causing undesirable behavior in applying secret/environment variables.
Suppose the user has an environment variable with filter `Git Repository= *node`. When he is creating a codespace with seed `Microsoft/vscode-remote-try-node` he will expect that the environment variable will be applied. But today since the portal is appending .git at the end, repo URL will not end with `*node` but with `*node.git`, hence the env var won't get applied.

Related work items: #1141487