Merged PR 255149: Add Diagnostics Server
This PR adds a "Diagnostics Server" used to improve the inner dev loop. The biggest features so far are live log parsing from the front and back end web services that show on a grid dashboard.

You should be able to run this by following the `README.md` in the Diagnostics folder.