Merged PR 256174: feat: emitting hotpath metrics via Mdm
This pull request contains conde to emit metrics via hotpath. Latency is expected to be around 60s.
We are going to calculate the success rate and latency.

The Mdm metric code related code is placed inside Vs saas sdk. So that multiple teams can consume this and emit metrics based on settings.