Merged PR 257288: [Diagnostics Server] Have FileLogWorker detect directory changes automatically
When running the Diagnostics server, it will scan the folders in `logs` for the different projects to monitor, and then scan for the newest log to process. However, it was only set up to automatically scan for new log files _within_ those folders. It would only scan for the folders themselves on first load. As a result, if you have never run the service before, you could end up seeing no logs being parsed since the service assumed log folders would exist already.

This PR fixes that by adding a file watcher for the log folder itself. If a new folder gets created or changed, it will start the scan over and add the new directories.