Merged PR 257488: [Diagnostics Server] Fix filter fields and bad react keys
There's a bad react key for the Log Table Filter fields in the settings popup.

When you touch or edit those specific filter fields, it's generating a new "key"... but when you edit the field, it causes a render since it's also changing the properties in the other elements of the grid.

This PR fixes that by using the `map` index when generating those fields, so it won't regenerate and won't lose focus.

Thanks to @<Rodrigo Moya Piernavieja> for pointing this out!