Merged PR 255838: fix: dev-stg environment fix to not have euw region
After adding euw region in dev-ci, dev-stg was broken since it was keep looking for euw region due to app settings override.

I have created a base 'appsettings.dev.json' and overriding it with dev-ci/dev-stg now to fix this issue as we really don't want to have another region in dev-stg due to less usage.