Merged PR 257696: Billing will self correct when missing events.
Billing can get lost if the last update > 48 hours. This has happened a few times and we also have had a few instances of underbilling. This change corrects those.
To fix it, billing now pulls all environment change records for each plan on each bill. It uses this to look for missing environments that may not have been in any past billing summary.

While this increases our RU usage a bit, it reduces it slightly in a different way. Currently, for each environment, there's a self correction check during calculations. This change pulls those potential multiple queries into a single top level query that pulls multiple environments.

With this change RU usage goes up for plans with a lot of older deader environments but drops it for plans with multiple active/suspended environments.

This additional RU cost will go away with the new billing strategy that we hope to implement before Sept.