Merged PR 259941: [FrontEnd] Add support for plan GET requests from RPaaS without user credential
While investigating a lot of 401 responses in the FE service for plan GET requests, we learned that RPaaS was making plan GET requests without a user credential which were failing because that credential is required by the auth middleware.

These requests were coming from an RPaaS background worker (thus the lack of user credential).  They were being made in an effort by RPaaS to keep caches in sync - anytime a plan GET request from a user returned a 404, RPaaS spawns this background process to verify that the resource is actually deleted.  If this background task also gets a 404 then RPaaS removes the resource from their cache (and thus deletes the resource from ARM).

We don't want this deletion to happen because mismatches of api-versions with dev/ppe can easily cause a 404 even if the plan actually exists.  RPaaS additionally _always_ uses the prod api version which is incorrect for dev/ppe plans.  The combination of these would cause a lot of incorrect deletes.

This PR adds support in the RPaaS auth middleware to make the user credential optional.  Currently only the GET endpoint makes use of this and in that cause _always_ returns a 200 without checking for the plan's existence.