Merged PR 259973: Rename environment access scopes
Rename scopes from "environments" to "codespaces". For compatibility we'll continue to validate the old scope names as being equivalent to the new names. This change will allow clients (and partners) to switch to using the updated scope names at any time.

I am not yet updating the code that issues tokens to use the scope names, to avoid issues with cross-location token use during rollout of the change. The environments controller endpoints check the scopes _before_ redirecting to the correct location.