Merged PR 259360: Move codespaces grid to overview page
This moves the codespaces grid onto the overview page and hides the tiles that were there. The whole transition is definitely a work in progress so I have kept the create codespace tile, which we will want to show when the user has no codespaces. CodespacesBlade.ts is also still in there for reference for moving the buttons.