Merged PR 260437: [workbench] Fix de-iframing issue for Safari
Tested locally with Safari.
This is a **workaround** for Safari similar to what we're using for Firefox already. On Safari we're having issues with indexedDB so use `InMemoryAsyncStorage ` instead.

This means the settings will be lost between sessions. The goal is to remove this workaround ASAP, which should happen with https://github.com/microsoft/vscs-planning/issues/191.