Merged PR 260691: Fix signalR service max packte size
The default value of the signalR service configuration is 32KB. When we use the relay hub to transport ssh messages this become a problem sice the encryption and other taxes added on the message transport will increase the packt size over 32KB.
We have the option to decrease the maximum values allowed on the ssh libraries (typescript and C#) but that also could exceed the max capacity when using the non messagepack option available in signalR clients which will base 64 encode big messages.
This PR will play safe and increment the max packat size to 64KB and so allow a safe room to later improve the ssh libraries (if we want).
Changes on the CLI will demo when you could increase this value and see how the server close the connection in that case,