Merged PR 257569: [workbench][auth][env]: add the ability for JWT auth middleware to use public certs from the HTTP endpoint
The biggest pitfall of sharing the portal source code with other teams is the fact that they need access to our key vaults.

This PR implements the `IsKeyvaultLess` mode and consumes the [recent SDK changes](https://devdiv.visualstudio.com/DefaultCollection/OnlineServices/_git/vssaas-sdk/pullrequest/257567) to add the ability for the JWT middleware to use local certs for validating the Cascade tokens.