Merged PR 261093: Fix x5c property in OIDC JWKS metadata
There was a bug in the `x5c` property of  the certificate info (JWKS) returned for OpenIDConnect metadata by the token service. This fix makes it conform to [the spec](https://auth0.com/docs/tokens/references/jwks-properties). I added a test case that uses the `OpenIdConnectConfigurationRetriever` to validate the metadata is well-formed.