Merged PR 261361: Switch tid claim in delegated plan access tokens
For the new `Microsoft.Codespaces` RP namespace, use the tenant ID of the calling user (that is the same as the plan subscription tenant ID) as the `tid` claim in delegated plan access tokens, instead of the plan resource ID guid.

Using this tenant ID is a better approach because it means delegated user IDs can be consistent across plans, rather than unique per-plan. As we are switching RP namespaces (and partners have not yet migrated), now is a good time to make this change.