Merged PR 262264: [PFS] PortForwarderController ExceptionView now uses correct redirect URL
Demo of the change: https://gfycat.com/frayedbiodegradableewe

The new `GetRedirectUrl` method uses code that was previously in the `SignIn` method. This code changes the redirect URL properly according to the request URL (e.g: if the URL is *.githubusercontent.com it redirects to `github.com/codespaces/auth` otherwise to our own endpoint).
This new method is now directly used by the `ExceptionView` code (in addition to `SignIn`).
With that, the `ExceptionView` code is lighter and can use the right `RedirectUrl`, which is then used in `exception.cshtml`.

Fixes https://github.com/github/codespaces/issues/753.