Merged PR 262330: Refresh status on Get should happen only when called from rest
Get request had a side effect that updates the environment state based on live share workspace status. Making that check to happen only when the GetAsync is invoked from the controller because liveshare API needs the user's cascade token to authenticate.
This will reduce the noise in the telemetry, that it's caused by the calls to the `EnvironmentManager.GetAync` from internal workers