Merged PR 263077: Connect scans for the appropriate Cascade Token if no PlanID supplied
## What's new?
If no PlanId is supplied to the connect page, we will scan all subs and all plans for the appropriate Cascade token.

We need this for the scenario in which the browser does not have the appropriate cookie and the user enters a URL to connect (ex. a port forwarding URL).

This is a temporary measure until we have a tenant-level API for mapping (ARM token, codespace ID) -> (plan ID)

## Limitations
The ajax() function in a React blade doesn't have any clear way of checking the response status (it works fine in a non-react blade) and the `fetch` API doesn't when running the extension locally (due to CORS limitations). So it is possible that we receive a 429 or something but fail to retry. However, a 429 from `/environments/{id}` seems unlikely because the limit is in the 100's/minute.