Merged PR 263731: Save tenant ID when creating plan
To support future functionality it will be helpful to know the tenant that a plan is in, so that we can direct a user to the correct tenant to log into to access the plan (in case they have only an environment ID but no other context). This change gets the tenant ID from the user's token when creating a plan, and stores it in our internal plan record. That always matches the tenant the plan's subscription is in, because the user has to use a token for the same tenant as the subscription in order to do ARM operations.