Merged PR 264524: Fix ARM user ID context
The user context set from the ARM user token was broken by recent refactoring in !258411. After that change the  current `HttpContext.User` identity is ignored if it is not an instance of `VsoClaimsIdentity` class. The fix here in the ARM user authentication handler is to wrap the identity in that class.

(I don't know if there are other places where this problem exists.)