Merged PR 262905: Job schedular API
This PR will allow register recurring or delayed job runs using a custom schedular.
The recurrign schedular wil use the 'cron' based expression based on the NCronTab library.
Note: Altough the ncrontab library has a nuget package the assembly is not signed and so rejected by the compiler, hence the need to get the source code on a separate project.
This PR will also define unit tests to cover the common scenarios and the integration with .NET Core DI.
Some helpers functions were added to also produce payload on the recurring or delayed job runs.