Merged PR 265085: Fix serialization of multi-value claims
When multiple values for a claim were provided, the values would get converted to a string containing a JSON array. Unfortunately the fix for that isn't that straightforward because of how .NET represents multiple claims having the same type as distinct `Claim` objects in a list. .NET must already be doing a similar conversion internally, but I could not find an API for it.