Merged PR 264683: Add Selection to Azure and GitHub Workbench Queries
Add a selection to the possible query params for github and azure workbenches.

This allows Rich Code Nav to generate direct links to sections of code in codespaces. Once this PR is merged, we will deploy [this PR](https://devdiv.visualstudio.com/DevDiv/_git/vscode-intelli-nav/pullrequest/264680) to add this functionality.