Merged PR 265256: ops: Initial core cluster template
ops: Initial core cluster template

This initial attempt uses a system-assigned managed identity instead
of a raw service principal for managing cluster resources. This is the
simplest approach and hopefully will work out. If not, we can continue
investigating user-assigned MI (BYO MI as the feature is called).

NOTE: It is necessary to specify the "System" mode on the agent pool in
order for this to deploy correctly.

The `TestCoreRegion.ps1` script is included to validate templates. It is
meant to be run from `Components.generated` so it can refer to a
generated template, not a source template.