Merged PR 263633: Add Nexus insertion to Sku catalog
This is needed for VS updates, as the VS team will insert the
update URL and version as they do the Windows image insertion,
so that the service (in future PRs) can use them to perform
VS updates on VMs.

This is on a separate PR so that the VS team can prepare their
insertions to include these new properties. On future PRs, code to
use this to send the VS update to the agent will be added.

See https://github.com/microsoft/vscs-planning/issues/21 for more details