Merged PR 266535: Moving Unavailable Environment Monitoring to Heartbeat Monitor.
Moving unavailable environment tracking.

I'm moving tracking for the unavailable state from the state transition monitor to the heatbeat monitor for a couple reasons:

- The state transition monitor is only kicked off when the transition first occurs, so turning it on now will not fix the environments already in that state
- The heartbeat monitor is currently reporting "healthy heartbeat" for these environments, which have been stuck in unavailable for days, just because a heartbeat is coming back at all, which seems strange. It should know to shut them off.

I still had to avoid the issue of shutting down those Windows environments though. I'm now avoiding the issue by tweaking the logic a bit. The environment must not only be in the unavailable state, but must have been there for the length of the timeout period (10min, 30 min, etc). The Windows issue causes the environments to flip between available and unavailable every second or two, so they would not get stuck in this logic. The state transition monitor only checks the current state at the end of the timeout period, so it's basically luck which state it catches those older Windows environments in.