Merged PR 266074: Support for Azure location on queue factory contracts
This PR is intented to support cross regions queues on the job queue API.
Some refactoring was needed in order to use the existing infra available.

The changes are backward compatible and now it is possible to reference a cross region queue to guarante the handler will be invoked on that region only.

Additional basic unit test is being provided for EUW region if the access keys are present.