Merged PR 265998: Enhance feature flag job scheduler usage by allowing runtime switching
In today implemention we have a feature flag that control the use of the job scheduler for the watch pool related jobs. This mechanism was only detected on startup and so it could cause potential problems durign re deployment or if we want to quickly switch to old/new type of task usage.
The changes were made with the perspective to add new feature flags when we start moving more code in the future.