Merged PR 264065: Add KMeans clustering based archival time predictions for environments
This PR adds dynamic computation of environment archival times. This mechanism can be turned on and off via a feature flag.

When an environment is suspended, an IEnvironmentArchivalTimeCalculator computes the archival time for an environment. This is saved in an Environment's ScheduledArchival property and gets used inside GetEnvironmentsReadyForArchiveAsync to query those environments that are ready for archival, falling back into the default cutoff time in case an environment has not being assigned a ScheduledArchival yet.

This PR also contains an implementation for IEnvironmentArchivalTimeCalculator (ClusteringBasedEnvironmentArchivalTimeCalculator) which uses a ML.NET KMeans clustering model to compute the archival time based on the behavior of environments, based on its BillingEvent history.

I completed end to end validations using a devstamp.