Merged PR 268345: Enable C# code analysis in vsclk-core
To meet an SDL requirement, I'm setting up the latest version of the Roslyn analyzers in the vsclk-core repo. Previously we had an old `CodeAnalysis.ruleset` file, but it had no effect since there was no analyzer package referenced by the build! The primary purpose of this PR is to get the latest security rules, but there are many other analyzer rules for reliability, maintainability, etc. Some of those others are just noise, but there are a number of rules that in my experience provide value, and are finding legitimate minor issues in our code.

- Reference the latest version of the Roslyn analyzers package (`Microsoft.CodeAnalysis.FxCopAnalyzers`)
- Set analyzer issues to be treated as build errors (for both debug and release builds).
- Enable all security rules.
  - Individual security rule violations can be suppressed in source on a case-by-case basis. In this PR I have suppressed a few, and fixed a few.
- Divide remaining non-security rules into recommended and non-recommended groups. There is no longer an official recommended list; there used to be one, but the needs of every project are different anyway. So I have used my own judgement here, based on past experience with code analysis across many codebases, and based on which rules were finding legitimate issues or reporting useless noise against the current code.
   - Disable all rules in the non-recommended list.
   - Enable rules in the recommended list that I can enable now without a lot of cleanup work.
   - We should clean up and enable the remaining recommended rules over time. I will create tasks to track that work.