Merged PR 271387: New continuation job handler framework
This PR will enable a new API for the continuation handler based on the job queues. In previous commits we define a new API to allow a job handler to be executed several times during a 'continuation' queue with retries and using the same job queue paradigm to allow much more flexibility on the concurrency.
On this PR we did make sure we use the API on a concrete implementation for the environment handlers. The approach is to allow the existing code modified as much as possible but sharing some helpers classes to avoid duplication. @ job handlers were implemented, the 'archive' and the 'start env' handlers. There is a new feature flags that will exercise this code.