Merged PR 273961: chore: logging the newly created workspaceid
chore: logging the newly created workspaceid

When a codespace is resumed, we have the old workspace id in the log, which isn't useful for tracking the resume progress.
This fixes it by logging NewWorkspaceId and NewSessionId right after it gets created, either in the continuation or during the request time.
We can track this workspace id later in the start/resume process to monitor the progress.