Merged PR 269500: Enable full PWA support for Chrome (& friends)
In order to make a PWA installable on Chrome (install button in URL bar) we needed the following changes:
1. Call the manifest file in workbench
2. Have png (specifically) icons of 192x192 and 512x512.

In addition, since we retired the old portal. The favicon has been updated to the GitHub/Codespaces icon.
Since Chrome loads the PWA install button icon as quickly as possible, we can't rely on javascript to dynamically change the favicon (that doesn't work).
Also the png icons are hardcoded in the manifest and this is required to meet the browser's PWA requirements.

Using the `fullscreen` option instead of `standalone` but it's not doing anything different for now in Chrome or Edge (eventually this should get rid of the top toolbar) and provide a 100% full screen experience like VSCode native.

For firefox, it looks like desktop doesn't support PWAs -> https://bugzilla.mozilla.org/show_bug.cgi?id=1407202
Same story for Safari, both browsers support that for mobile and we meet the requirements.

Note: see the web team demo and conversations for the favicon change.

Note 2: Unfortunately users cannot change the Codespace name and it has to be hardcoded in the manifest. I believe there are talks in Chromium to change that.

Pics or it didn't happen 👇
![Screen Shot 2020-08-21 at 4.08.53 PM.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/269500/attachments/Screen%20Shot%202020-08-21%20at%204.08.53%20PM.png)

![Screen Shot 2020-08-21 at 4.09.16 PM.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/269500/attachments/Screen%20Shot%202020-08-21%20at%204.09.16%20PM.png)

![Screen Shot 2020-08-21 at 4.09.28 PM.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/269500/attachments/Screen%20Shot%202020-08-21%20at%204.09.28%20PM.png)