Merged PR 269826: Allowed States for Suspend
https://github.com/microsoft/vssaas-planning/issues/472

Updated EnvironmentSuspendAction.cs to use a switch statement to check the current state of the environment, determine whether suspend should be allowed, and act accordingly.

One change from the allow list discussed in the github link above is that suspend is now allowed when the environment is in the Failed state. In order to delete a Failed environment, I found that the system suspends the environment and then deletes it. We want this suspend to succeed, so I allow suspend in Failed state here.

Another note is that EnvironmentForceSuspend is triggered when a suspend is called in the Awaiting, Unavailable, and Failed states. This reflects the suspend action's current workflow, which triggers force suspend when the environment is not in the Available state.