Merged PR 273446: Auto throttle support on job queues consumers
We want a better mechanism to control how the queued messages are being consumed by the job handlers. Today when a queue message is available on a particular queue id is being dequeued right away into the consumer without any specific limit on the queue capacity. This PR will allow better customization on how we dequeue messages.
The billing V2 startup code was changed to allow such customization and allow only 5k messages to be available, the rest available will sit on the Azure queue instead of being in memory.
Also we introduce the concept of job error handlers where we can share different patterns on how to process an exception and provide a mechanism on how to retry those jobs.