Merged PR 275033: Fixed up some VsoUtil commands
Fixed up some VsoUtil commands

Added a Comment field to SystemConfigurationRecord so that we don't lose any comment value
that exists when updating the configuration value.

The ToggleFeature command's --value argument needs to be a string so that we can parse
the value. A bool parameter isn't handled as expected.