Merged PR 276259: Return 421 if request ends up at the wrong cluster
Return 421 if request ends up at the wrong cluster

Fixes an issue in Firefox where it can re-use TLS connections for the wrong cluster due to the connection resuse logic in the HTTP/2 specification. 421 is the proper response according to the protocol to instruct the browser not to re-use that connection and to replay the request with a new connection, which will properly resolve DNS and route the request to the correct cluster/IP.

The scenario that we were seeing is that some requests to a region-specific endpoint, such as `westus2.online.visualstudio.com` were landing on the wrong regional cluster, which should never happen according to DNS. But due to a quick of the HTTP/2 connection reuse rules, this could happen. So now the frontend will specifically validate the Host header and ensure it is only processing requests for domains it expects: the global endpoint and it's own regional endpoint. Any requests for another regional endpoint that it shouldn't process will return 421.

Good request:
![localhost-success.PNG](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/276259/attachments/localhost-success.PNG)

Bad request:
![jokeech-localhost-fail.PNG](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/276259/attachments/jokeech-localhost-fail.PNG)