Merged PR 276202: Fix bounded capacity for job handlers
After analyzing ppe telemetry find out several surprises around the max value of the input count on some of the billing V2 job handlers. I was expecting them to be always lower than 32 but turn out the value get bigger than that.
After looking on the code find out that the lack of bounded capacity on the job handler itself make consumption of the message queue from Azure Queue pointless since it will always find a way to buffer the items being collected.

That make one of the setting of the queue message producer not usable for the needed purpose. The solution proposed require to move most of the code from the queue message producer back to the job consumer so he can always produce jobs directly and pass it job handler action block directly. The new way to control the input buffer is by limiting the 'bounded capacity' of the custom job handler, that gives us full control on what we want to achieve with the best perf possible since we avoid the extra TPL entities but by having directly the action block tied to the queue message producer.