Merged PR 275665: Migrate LogSystemResourceStateTask to use Job Queue API and Job Scheduler API
This PR is to migrate LogSystemResourceStateTask to use our new Job Queue API and Job Scheduler API.

We separated the code into producer and consumer. LogSystemResourceStateJobProducer is responsible for creating the LogSystemResourceStatePayload. The payload is consumed by LogSystemResourceStateJobHandler.

The old LogSystemResourceStateTask will be replaced by new producer and handler. No setting flag is needed.