Merged PR 276807: Bypass SKU Validation for GH Codespaces
## Why is this change needed?
GH will maintain it's own list of SKUs and will be the authority on which users can create which SKUs. For cases like our private preview of Windows SKUs, we don't want to have to update both the GH list and our list, so we simply bypass the visibility check if the request is coming from GH.

## Won't we need to update the list SKUs api for GH?
Because GH maintains its own list of SKUs, such a change is not required.

## How was this tested?
I wasn't able to fully test this because I don't think we have a way to send GH requests to our machine. However, I tried a couple things to validate what I could:
1) F5'd the FES locally and created a standardLinux Codespace via the extension to ensure no regressions
2) I did step 1 again, but with a breakpoint at the beginning of `ValidateSkuAsync`. At that point, I manually changed the partner to "GitHub" and they skuName to "internalWindows". I ensured SkuUtils.IsVisible was never called and the environment was created successfully.