Merged PR 276963: [portal] Add request beta access link only when logged in
Fixes: https://github.com/microsoft/vssaas-planning/issues/696

Description of the change
--------------------------

When logged in, users should see a "Request beta access" link. This points them to a survey which, once filled, allows us to add them to GitHub Codespaces.
This code checks for the `CachedToken` to ensure the user is logged in. Therefore we won't show the survey link on the login page.

Impact
-------

Important fix to transition VSO users to GitHub. Otherwise we're relying on them using VSCode desktop (toast notification available there) or us reaching out to them.

Rollback Plan
-------------

This fix can be reverted without causing any issue