Merged PR 276731: Migrate WatchOrphanedPoolTask to use Job Queue API and Job Scheduler API
This PR is to migrate WatchOrphanedPoolTask to use our new Job Queue API and Job Scheduler API.

We separated the code into producer and consumer. WatchOrphanedPoolJobProducer is responsible for checking pools in the system and creating WatchOrphanedPoolPayload for the pool that needs to be deleted.

The payload is consumed by WatchOrphanedPoolJobHandler. The WatchOrphanedPoolJobHandler will perform the deletion operation.

A setting flag "setting:vsclk-usw2:watchorphanedpooljob-enabled" is required in backend Cosmos DB configuration for this migration.

This PR also removes the deprecated LogSystemResourceStateTask.