Merged PR 276866: Migrate continuation job handlers on the back end
The proposed PR will migrate the following continuation job handlers:
1. DeleteResourceContinuationJobHandler
2. ResumeEnvironmentContinuationJobHandler
3. ExportEnvironmentContinuationJobHandler
4. ResourceHeartbeatJobHandler

To achieve this we did create a generic class on our 'Common' assembly that is based on the original env manager continuation job handler. This class will have specific implementation for the FE and the BE.
Please note we leave untouched the original classes but we create a deprecated folder where we expect to delete once the feature flag prove stability.