Merged PR 277427: [ops] Override-enable more dev subs
Description of the Change
-------------------------

Update the `dbEnabled` property on more new dev subscriptions, so they can be turned on using the override setting in the backend database.

Note that two of these (`vscs-codesp-dev-data-network-eu-w-001` and `vscs-codesp-dev-data-network-us-w2-001` were already manually overridden in the database, but `subscriptions.json` was not previously updated to reflect that).

Impact
------

Not critical.

Rollback Plan
-------------

Reverting alone is not sufficient to rollback. Would have to also run this pipeline after the revert merged: https://dev.azure.com/devdiv/OnlineServices/_releaseDefinition?definitionId=289&_a=definition-pipeline .