Merged PR 277803: Fix two issues with export
Fixes: https://github.com/microsoft/vssaas-planning/issues/877

Description of the Change
-------------------------

Fix two issues with export

1) Exporting was broken due to an invalid cast when starting the compute
2) When exporting fails, it tries to "revert" the export state change by triggering a suspend action, but the suspend action did not handle the exporting state. Updated the suspend action to allow suspension from exporting.

I also renamed the parameter in the ResourceStartManager to better align with its type and to prevent confusion about the expected type of the data coming in.

Impact
------

Currently the export action is broken. This change should allow export to work.

Rollback Plan
-------------

This change is safe to revert.