Merged PR 276511: Fix plan tenant migration
 - When migrating a codespace to a plan and that plan record doesn't already have a tenant, use the tenant from the destination plan access token, not the source plan access token.
 - Improve logging of plan and tenant info during migration

Background: For newly created plans, we record the tenant ID with the plan record when creating the plan, and that would be used when migrating into the plan. But there was a short period of time when the new RP was enabled but we did not yet record the tenant of plans created in the new RP. This fixes migration of codespaces into those plans.