Merged PR 277975: feat: adding printInline command for kusto tool.
Fixes: [link to issue]

Description of the Change
-------------------------

Adding printInline functionality in the internal kusto tool.

Impact
------

It is an internal tool. No impact to code as-is.

Rollback Plan
-------------

Yes, this can be rolled back.

feat: adding printInline command for kusto tool.

All dependent functions are printed inline. It can be used for testing locally without modifying the stored functions in the db.
Plus some fixes.